feat: add optimistic delete of expenses
